FEATURE: Implemented Blurry Sidebar
Edited the sidebar to include a blurry background that changes as the camera view changes. Furthermore, as a part of the new improved blurry sidebar, the table view was made custom inside of it. Gestures were implemented to open and close the sidebar in addition to a single tap gesture.

Resolves: Inability to close sidebar with tap, poorly looking sidebar