FEATURE: Ability to create Knots from XML files
The app is now able to create Knot objects from an XML file so that it is an easy process to modify and create new knots. The XML file contains the name of the knot, the name of the thumbnail of the knot, and knot states. This essentially maps the different knot states and animations from their files to the Knot object, so that SCNNodes can switch on demand.

Resolves: clean and easy way to link files with their intended content