FEATURE: KnotStateEditor play bar
This commit adds a bar on the bottom that can be used to adjust the KnotState or the portion within the knot tutorial that is displayed on AR. This includes a UISlider which can be used to slide through the tutorial, or forward and back buttons to move the slider along automatically. It calls a delegate reponsible for adding the appropriate SCNNode.

Resolves: Inability to move through the tutorial